requirements.txt
fastapi
uvicorn
beanie
motor
python-dotenv
httpx
pytest
pytest-asyncio